agent/requirements.txt
fastapi
uvicorn
google-generativeai
pyyaml
python-dotenv
sqlite3
